Added support for actions to the dashboard.
The configs now have "actions" so they know what to do when a widget is in a given state. Then, wrote some logic so that when a device widget is clicked/tapped, it will send the request to ST to perform the action.

Also improved the efficiency of updates to HTML of the widget.