Major update makes 1 win that handles all screens.
The index page is the loader page, and it loads in all other screens.

It starts with a splash screen that doesn't clear until the dashboard is done loading.

Notification for contact sensors was rewritten to appear on top of the dashboard screen.

The animate.css library was used to make effects real nifty. Also, I made a logo. :-D