Changes to ST app to support Intrusion notification.
It doesn't actually use the official intrusion notifications of SHM, rather, it uses a switch or an alarm device (if you don't have one, you can make a virtual "simulated" one) that you set to get turned on when SHM detects an intrusion.  The RPi code will turn off the notifier switch/alarm but not the actual SHM intrusion status.