Wire up the keypad code to work with intrusions.
So it is wired up to the rest of the app, but also enhanced to do necessary listeners and AJAX calls for notification back and forth.

Support for any code to have a URL that is executed. Can be used for a distress code to execute a URL but doesn't have to be exclusively for distress codes.

Presently the countdown while waiting for the key code will trigger the alarm on the client side. Eventually the triggering of the alarm (and distress actions for non-action) should be done server side in case equipment is turned off or destroyed.