Updated to use state data to store client URI.
Previously, it was housed in the preferences, but the custom text field for the client URI does not carry over nicely in the web OAUTH screen. So instead, I'll make the client send its URI to the app via a web service call.

So this has web service calls to get, add, and delete client URIs.