Improve SmartThings app to remove stale client URIs.
If a client URI is coming back with errors, suggesting it's not online anymore, then disconnect from it after 20 failures.

The state.uri variable previously was an array of strings which were the URI strings. Now it is an array of objects, which hold the URI and how many failures there have been. When there is a success, the failure count is reset.